feat: api code refactor
