+++ src/util/dbConnection.ts
@@ -1,0 +1,13 @@
+import mongoose, { ConnectOptions } from "mongoose";
+
+export async function connectToMongoDB() {
+  try {
+    await mongoose.connect(`${process.env.MONGODB_URI}`, {
+      useNewUrlParser: true,
+      useUnifiedTopology: true,
+    } as ConnectOptions);
+    console.log("Successfully Connected to MongoDB!");
+  } catch (error) {
+    console.error("Error connecting to MongoDB:", error);
+  }
+}
